requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
scikit-learn>=1.3
altair>=5.2
vegafusion>=1.6
vl-convert-python>=1.6
joblib>=1.3
